Major refactor of the entire codebase to use ImmutableJS 100% of the time.
Previously, some parts of the system(building search) would pass in a
pojo, while the primary contractor interface would pass in an immutable
object.  But then sometimes inner components would convert the immutable
to a pojo, and attempt to do that in getDerivedStateFromProps.  This was
extremely fragile, and was the root cause of the immense slowdown
whenever a building was clicked on the map.

After this change, immutable instances are used everywhere.  Redraws are
nearly instantaneous.  Building search is faster, but not yet where it
needs to be.  That will come later.

Additionally, react-immutable-proptypes is now in use in a large
percentage of the code-base, but Canvas and GISSearch still needs a bit
of work.